Update LICENSE
All source files have headers designating them as Apache-2.0-licensed.
The `LICENSE` file however contains the text of the LGPL 3. This causes
e.g. Github to declare the project as being LGPL licensed and also
causes problems for packagers who need to figure out the project's
license.

Since the `LICENSE` file seems to be a remnant of a distant past, update
it to contain the Apache license, which seems to be what the project is
intended to be licensed under.

Signed-off-by: Conrad Hoffmann <ch@bitfehler.net>