feat: Make handling of invalid devices configurable
Currently, whenever there are devices that are invalid, smartctl_exporter will
log an error stating that the device can not be opened, refering to potential
issues with missing IDENTIFY_DEVICE structures or low-power-mode-devices.

There are, however, also situations where (virtual) devices disappear
from the host and are not properly cleaned up. Here, `open()`ing the
device in smartctl fails with ENXIO, resulting in the error message
"No such device or address" being returned. In this case, the error
just clutters the log output for no reason, as there is no reasonable
action that can be taken.

This MR makes the experienced behavior for such cases configurable.
One can either specify the newly added flag `--smartctl.exclude-enxio`
to ensure that devices with such problems are simply silently ignored,
or decide to ignore the flag and continue experiencing the current
behavior.

Signed-off-by: Markus Kostrzewski <markus.kostrzewski@hetzner.com>